refactor: 独立干员卡片组件
